+++ RELEASE_NOTES.md
@@ -11,7 +11,10 @@ Version 0.2.1 (November 26th 2016)
   * Added copyScaledVectorTo() method merging copy and scale methods into one
   * Added copySubtractFromTo() method merging copy and subtract methods into one
   * Added isPositive() and isNegative methods
-  * Added bottomRightCorner and halfSize properties to rectangle geometry class
+  * Added setXToOrigin() and setYToOrigin() methods
+ * Geometry class
+  * Added bottomRightCorner and halfSize properties to rectangle class with getters and setters
+  * Added size and halfSize Properties to circle class
 
 Version 0.2.0 (September 21st 2016)
 ------------------------------